refactored patterns to be parsed only once and both patterns and parsing in seperate file
